1. fix board mod 2. fix board protype 3. vip read updated
