+++ src/components/adi/common/board/Board.vue
@@ -4,65 +4,6 @@ import compBaseMixin from '../comp.base.mixin'
 import boardProptypes from './board.proptypes'
 import { mapGetters, mapActions } from 'vuex'
 
-let isInitEditor = false
-
-const initEditor = (data, callback) => {
-  isInitEditor = true
-
-  let mxClientEle = document.querySelector('#mx-client')
-
-  if (!window.mxClient.isBrowserSupported()) {
-    mxClientEle.innerHTML('Browser is not supported!')
-  }
-
-  let mxClientHeight = window.innerHeight
-  let mxClientWidth = window.innerWidth
-
-  mxClientEle.style.width = mxClientWidth + 'px'
-  mxClientEle.style.height = mxClientHeight + 'px'
-
-  window.mxResources.loadDefaultBundle = false
-
-  let bundle =
-    window.mxResources.getDefaultBundle(RESOURCE_BASE, mxLanguage) ||
-    window.mxResources.getSpecialBundle(RESOURCE_BASE, mxLanguage)
-
-  window.mxUtils.getAll(
-    [bundle, window.STYLE_PATH + '/default.xml'],
-    function(xhr) {
-      window.mxResources.parse(xhr[0].getText())
-
-      let themes = new Object()
-      themes[
-        window.Graph.prototype.defaultThemeName
-      ] = xhr[1].getDocumentElement()
-
-      let ui = new window.Board(
-        new window.Editor(urlParams['chrome'] == '0', themes),
-        mxClientEle
-      )
-
-      if (
-        data &&
-        data.replace(/[\ \r\n]+/g, '').length > 0 &&
-        data.replace(/[\ \r\n]+/g, '') != 'blank'
-      ) {
-        let doc = ui.editor.graph.getDecompressData(data)
-
-        ui.editor.setGraphXml(doc.documentElement)
-      }
-
-      if (typeof callback == 'function') {
-        callback(ui)
-      }
-    },
-    function() {
-      mxClientEle.innerHTML =
-        '<center style="margin-top:10%;">Error loading resource files. Please check browser console.</center>'
-    }
-  )
-}
-
 const initPreview = (data, callback) => {
   if (!window.mxClient.isBrowserSupported()) {
     return 'Browser is not supported!'
@@ -146,18 +87,6 @@ export default {
             return <div class="mx-client-start">点击右侧编辑按钮开始编辑</div>
           }
         })()}
-        <el-dialog {...this.getDialogProps}>
-          <div id="mx-client">
-            <div
-              class="mx-client-close"
-              on-click={() => {
-                this.closeDialog()
-              }}
-            >
-              关闭
-            </div>
-          </div>
-        </el-dialog>
       </div>
     )
   },
@@ -169,46 +98,6 @@ export default {
   },
   mixins: [compBaseMixin],
   methods: {
-    ...mapActions({
-      setActiveProperty: 'setActiveProperty',
-      setActivePropertyData: 'setActivePropertyData'
-    }),
-    ...mapGetters({
-      activeProperty: 'activeProperty'
-    }),
-    showDialog() {
-      this.visible = true
-    },
-    closeDialog() {
-      this.visible = false
-
-      let data = this.ui.getCurrentCompressData()
-
-      if (this.activeProperty()) {
-        this.changeProptyData({ data })
-      }
-
-      let mxWindow = document.querySelectorAll('.mxWindow')
-
-      if (mxWindow) {
-        mxWindow.forEach(element => {
-          element.parentNode.removeChild(element)
-        })
-      }
-    },
-    loadBoardEditor() {
-      if (!isInitEditor) {
-        initEditor(this.properties.data, ui => {
-          this.ui = ui
-        })
-      }
-    },
-    changeProptyData(changedData) {
-      // this.changeActivePropty()
-      this.setActivePropertyData({
-        data: changedData
-      })
-    },
     sync(prop, value) {
       this[prop] = value
     }
@@ -235,22 +124,26 @@ export default {
   },
   created() {
     //TODO: Rewrite mxGraph to es6 code
-    let boardScript = document.createElement('script')
-    boardScript.setAttribute('src', './static/adi/board/keepwork-board.min.js')
 
-    let graphEditorCss = document.createElement('link')
-    graphEditorCss.setAttribute('rel', 'stylesheet')
-    graphEditorCss.setAttribute('type', 'text/css')
-    graphEditorCss.setAttribute(
-      'href',
-      './static/adi/board/assets/styles/grapheditor.css'
-    )
+    if (!window.mxClient) {
+      let boardScript = document.createElement('script')
+      boardScript.setAttribute(
+        'src',
+        './static/adi/board/keepwork-board.min.js'
+      )
 
-    let body = document.querySelector('body')
-    body.appendChild(boardScript)
-    body.appendChild(graphEditorCss)
+      let graphEditorCss = document.createElement('link')
+      graphEditorCss.setAttribute('rel', 'stylesheet')
+      graphEditorCss.setAttribute('type', 'text/css')
+      graphEditorCss.setAttribute(
+        'href',
+        './static/adi/board/assets/styles/grapheditor.css'
+      )
 
-    // boardProptypes.click = this.showDialog
+      let body = document.querySelector('body')
+      body.appendChild(boardScript)
+      body.appendChild(graphEditorCss)
+    }
   }
 }
 </script>
@@ -264,35 +157,6 @@ export default {
 
 <style lang="scss" scoped>
 .comp-board {
-  .mx-client-close {
-    position: absolute;
-    right: 0;
-    top: 0px;
-    z-index: 9999;
-    -webkit-box-shadow: none;
-    -moz-box-shadow: none;
-    box-shadow: none;
-    background-color: #4d90fe;
-    background-image: linear-gradient(top, #4d90fe, #4787ed);
-    border: 1px solid #3079ed;
-    color: #fff;
-    border-radius: 2px;
-    cursor: default;
-    font-size: 11px;
-    font-weight: bold;
-    text-align: center;
-    white-space: nowrap;
-    margin-right: 16px;
-    height: 27px;
-    line-height: 27px;
-    min-width: 54px;
-    outline: 0px;
-    width: 90px;
-    padding: 0 8px;
-    padding: 0 8px;
-    cursor: pointer;
-  }
-
   .mx-client-start {
     background-color: #dedede;
     color: #333;
