+++ src/components/adi/common/board/board.properties.js
@@ -1,4 +1,3 @@
 export default {
-  name: '按钮',
   data: ''
 }
